Refactor getQueries so it's useful in other parts of store stats
